backend/requirements.txt
# FastAPI
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic[email]==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# AWS
boto3==1.34.34
botocore==1.34.34

# Azure
azure-identity==1.15.0
azure-mgmt-subscription==3.1.1
azure-mgmt-resource==23.0.1
azure-mgmt-storage==21.1.0
msgraph-sdk==1.2.0

# Security
cryptography==42.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6

# PDF Generation
reportlab==4.0.9
weasyprint==60.2

# Utils
python-dotenv==1.0.0
httpx==0.26.0
tenacity==8.2.3
aiofiles==23.2.1
pytz==2024.1

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
